metadata_extraction/requirements.txt
fastapi
uvicorn
pytesseract
mammoth
Pillow
transformers
datasets
pandas
